Add check for large collections in Liquid::For
If a large collection is given to a for block, this may crash the render
process of the template. Setting of resource_limits does not help
because these are tested at a later stage.

This can be prevented by adding a simple check to the Liquid::For block
which makes sure that a collection size does never exceed the set
render_score_limit.