Fix scrollbar and scrolling active link into view on the sidebar
Scrollbar now scrolls over to the actual link and positions it at the center of sidebar. This is less jarring that scrolling to the section header at the top. The logo is also moved into the scrollable sidebar but made sticky through CSS.

Tested working on both desktop and mobile.