Fix the anchor links in Introduction and Template
For Introduction, because the sidebar menu also contain the same anchors, they create a conflict for the anchor links in the first paragraph. The solution is to manually render the headings with a modified anchor name. (Objects header was included for consistency and future-proofing.)

For Template, because of the presence of the version badge, the anchor name has the stripped version number added to it. Hence the anchor links need to reflect that.