Add a description for ifchanged in Iteration page
Since this exists and it's currently undocumented, it should now be documented and explained in the documentation. If this will be declared as "deprecated", a deprecation tag can be added later (stating the version number at which this became deprecated).

Also, according to the original source file ifchanged.rb the logic is to check the output generated by the code between ifchanged and endifchanged, and compare it to the previous output. The logic is NOT to check whether the loop variable has changed. Even if the loop variable remained the same but there was some other change in the output (for example, using cycle tag), the ifchanged tag will print the output.